pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "privacy-pipeline"
version = "0.1.0"
requires-python = "~=3.10"
dependencies = [
    "datasets==2.17.1",
    "fastapi==0.110.3",
    "hydra-core==1.3.2",
    "openai==1.12.0",
    "pydantic==2.6.2",
    "rouge-score==0.1.2",
    "starlette==0.37.2",
    "torch==2.2.1",
    "transformers==4.40.1",
    "httpx==0.27.0",
    "gritlm==1.0.2",
    "faiss-cpu==1.8.0",
    "numpy==1.25.2",
    "pyserini",
    "tiktoken==0.7.0",

]

[tool.hatch.build.targets.wheel]
packages = ["src"]
